pyproject.toml
[tool.poetry]
name = "dns-shop-backend"
version = "0.1.0"
description = ""
authors = ["Alexey Oleynik <a20051772@yandex.ru>"]
packages = [{include = "app"}]


[tool.black]
line-length = 160
skip-string-normalization = true
include = '\.py$'
exclude = '''
(
    /(
        \.eggs
      | \.git
      | \.mypy_cache
      | deploy
      | \tests
      | \.venv
      | \sql
    )/
  | /(
        __pycache__
      | _mypy_cache
    )/
)'''


[tool.mypy]
ignore_missing_imports = true
disallow_untyped_calls = false
check_untyped_defs = true
warn_redundant_casts = false
show_error_codes = true
pretty = true
ignore_errors = false
exclude = 'tests'


[tool.isort]
profile = "black"
line_length = 160
multi_line_output = 3
lines_after_imports = 2
lines_between_sections = 1
known_local_folder  = ['app']
known_first_party = ['fastapi', 'starlette', 'pydantic', 'pytest', 'uvicorn', 'httpx', 'tortoise']
sections = ['FUTURE', 'STDLIB', 'FIRSTPARTY', 'THIRDPARTY', 'LOCALFOLDER']

[tool.semgrep]
additional_args = "--severity=error"
paths = ["app/"]
exclude_paths = [".venv","./app/tests/*.py"]


[tool.poetry.dependencies]
python = "~3.12"
httpx = "~0.25.0"
fastapi = "~0.103.2"
pyyaml = "~6.0"
uvicorn = {extras = ["standart"], version = "~0.23.2"}
gunicorn = "~21.2.0"
#aerich = {extras = ["asyncpg"], version = "^0.7.2"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
python-jose = "^3.3.0"
bcrypt = "^4.1.1"
passlib = "^1.7.4"

[tool.poetry.group.test]
optional=true

[tool.poetry.group.test.dependencies]
pytest = "~7.4.2"
pytest-mock = "~3.11.1"
pytest-asyncio = "~0.21.1"

[tool.poetry.group.dev]
optional=true

[tool.poetry.group.dev.dependencies]
flake8 = "~6.1.0"
black = "~23.9.1"
pre-commit = "~3.4.0"
mkdocs = "~1.5.3"
mkdocs-material = "~9.4.4"
coverage = "~7.3.2"
flake8-quotes = "~3.3.2"
gitlint = "~0.19.1"
flake8-print = "~5.0.0"
isort = "~5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "app"

#[tool.aerich]
#tortoise_orm = "internal.repositories.db.helpers.TORTOISE_ORM_CONFIG"
#location = "app/internal/repositories/db/migrations"
#src_folder = "./app"